[fixed] Ship Subcomponents populating fields that they do not change.

+++ MMOCoreORB/src/server/zone/objects/ship/components/ShipShieldComponentImplementation.cpp
@@ -47,15 +47,29 @@ void ShipShieldComponentImplementation::fillAttributeList(AttributeListMessage* 
 
 	StringBuffer msg;
 
-	msg << "0.0/" << Math::getPrecision(frontHitpoints, 1);
-	alm->insertAttribute("@obj_attr_n:ship_component.ship_component_shield_hitpoints_front", msg.toString());
-	msg.deleteAll();
+	if (getGameObjectType()==SceneObjectType::SHIPSHIELDSUBCOMPONET){
+		alm->insertAttribute("@obj_attr_n:ship_component.ship_component_shield_hitpoints_front", String::valueOf(Math::getPrecision(frontHitpoints, 1)));
+	}else {
+		if (frontHitpoints != 0.f) {
+		msg << "0.0/" << Math::getPrecision(frontHitpoints, 1);
+		alm->insertAttribute("@obj_attr_n:ship_component.ship_component_shield_hitpoints_front", msg.toString());
+		msg.deleteAll();
+		}
+	}
 
-	msg << "0.0/" << Math::getPrecision(rearHitpoints, 1);
-	alm->insertAttribute("@obj_attr_n:ship_component.ship_component_shield_hitpoints_back", msg.toString());
-	msg.deleteAll();
+	if (getGameObjectType()==SceneObjectType::SHIPSHIELDSUBCOMPONET){
+		alm->insertAttribute("@obj_attr_n:ship_component.ship_component_shield_hitpoints_back", String::valueOf(Math::getPrecision(rearHitpoints, 1)));
+	}else {
+		if (rearHitpoints != 0.f) {
+		msg << "0.0/" << Math::getPrecision(rearHitpoints, 1);
+		alm->insertAttribute("@obj_attr_n:ship_component.ship_component_shield_hitpoints_back", msg.toString());
+		msg.deleteAll();
+		}
+	}
 
-	alm->insertAttribute("@obj_attr_n:ship_component.ship_component_shield_recharge_rate", String::valueOf(Math::getPrecision(rechargeRate, 1)));
+	if (rechargeRate != 0.f) {
+		alm->insertAttribute("@obj_attr_n:ship_component.ship_component_shield_recharge_rate", String::valueOf(Math::getPrecision(rechargeRate, 1)));
+	}
 
 	if (reverseEngineeringLevel != 0) {
 		alm->insertAttribute("@obj_attr_n:reverseengineeringlevel", reverseEngineeringLevel);
